Changing library from ART to react-native-svg
This change removes ART and adds react-native-svg as the drawing library for producing the barcode. This change is to support the deprecation of ART from the ReactNative libraries. The react community ART library is not linked in Expo v36 projects. SVG is the preferred method. This change is linked to issue #49 https://github.com/wonsikin/react-native-barcode-builder/issues/49